requirements.txt
# Telegram Mini App - Requirements
# Python 3.10+

# Production server (опционально, для production используйте Gunicorn)
# gunicorn>=23.0.0

# Telegram Bot
aiogram>=3.13.1

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# Database
aiosqlite>=0.20.0

# File handling
aiofiles>=24.1.0

# HTTP client
httpx>=0.27.0

# Validation & Settings
pydantic>=2.4.1,<2.9
pydantic-settings>=2.0.0

# Security
python-jose[cryptography]>=3.3.0

# Environment variables
python-dotenv>=1.0.1

